# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GeneSilico EHR API",
        "summary": "Ontada (iKnowMed) FHIR + Stedi eligibility for Texas Oncology.",
...
        "contact": {
            "name": "GeneSilico",
            "email": "developers@genesilico.ai"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8090",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/preauth": {
            "get": {
                "tags": [
                    "preauth"
                ],
                "summary": "List Packages",
                "operationId": "list_packages_api_preauth_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackageList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "preauth"
                ],
                "summary": "Create Package",
                "operationId": "create_package_api_preauth_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackageOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/preauth/{package_id}": {
            "get": {
                "tags": [
                    "preauth"
                ],
                "summary": "Get Package",
                "operationId": "get_package_api_preauth__package_id__get",
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Package Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackageOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/preauth/{package_id}/lines": {
            "post": {
                "tags": [
                    "preauth"
                ],
                "summary": "Set Lines",
                "description": "Replace the billable lines \u2014 this is how a drug tier gets assigned.",
                "operationId": "set_lines_api_preauth__package_id__lines_post",
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Package Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LinesIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackageOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/preauth/{package_id}/benefits": {
            "get": {
                "tags": [
                    "preauth"
                ],
                "summary": "Package Benefits",
                "description": "Parsed benefits from the 271 already stored on this package.",
                "operationId": "package_benefits_api_preauth__package_id__benefits_get",
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Package Id"
                        }
                    },
                    {
                        "name": "network",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "in",
                            "title": "Network"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BenefitsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/preauth/{package_id}/submit": {
            "post": {
                "tags": [
                    "preauth"
                ],
                "summary": "Submit",
                "description": "Deliberately unimplemented.\n\nSubmitting a prior authorisation is an X12 278 transaction. Stedi exposes\nno 278 endpoint \u2014 verified against both the test and the production key,\nevery candidate path returns 404. Returning a fabricated authorisation\nnumber here would be the single most damaging thing this service could do,\nso it refuses instead.",
                "operationId": "submit_api_preauth__package_id__submit_post",
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Package Id"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Always. There is no X12 278 channel.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BenefitsResponse": {
                "properties": {
                    "benefits": {
                        "$ref": "#/components/schemas/ParsedBenefits"
                    },
                    "estimate": {
                        "$ref": "#/components/schemas/Estimate"
                    },
                    "source": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Source"
                    }
                },
                "type": "object",
                "required": [
                    "benefits",
                    "estimate",
                    "source"
                ],
                "title": "BenefitsResponse"
            },
...
            "ErrorEnvelope": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorEnvelope",
                "description": "Every non-2xx response from this API has exactly this shape."
            },
...
            "LinesIn": {
                "properties": {
                    "lines": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Lines"
                    }
                },
                "type": "object",
                "required": [
                    "lines"
                ],
                "title": "LinesIn"
            },
...
            "PackageIn": {
                "properties": {
                    "casebook_id": {
                        "type": "string",
                        "title": "Casebook Id"
                    },
                    "regimen": {
                        "type": "string",
                        "title": "Regimen",
                        "default": ""
                    },
                    "payer_name": {
                        "type": "string",
                        "title": "Payer Name",
                        "default": ""
                    },
                    "member_id": {
                        "type": "string",
                        "title": "Member Id",
                        "default": ""
                    },
                    "group_number": {
                        "type": "string",
                        "title": "Group Number",
                        "default": ""
                    },
                    "hospital_id": {
                        "type": "string",
                        "title": "Hospital Id",
                        "default": ""
                    },
                    "lines": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Lines"
                    }
                },
                "type": "object",
                "required": [
                    "casebook_id"
                ],
                "title": "PackageIn"
            },
            "PackageList": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/PackageOut"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "title": "PackageList"
            },
...
        }
    },
...
}